backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
alembic==1.13.2
python-dotenv==1.0.1
redis==5.0.7
rq==1.16.1
clickhouse-driver==0.2.9
passlib[argon2]==1.7.4
python-jose==3.3.0
pydantic[email]==2.8.2
pydantic-settings==2.4.0
prometheus-client==0.21.0
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.23.8
psycopg2-binary==2.9.9
